solver/requirements.txt
websocket-client==1.0.1
